Organize 4.0.0 change list
Keeping all changes in a single list means users only have to look
in one place when they upgrade.

In this case, the "Removed" item was really obvious, but in general
a single list is a good practice, I think.

A single list is also recommended by http://keepachangelog.com/
which I'm a big fan of.

[ci skip]