Use updated_at timestamp for "create" versions
"Create" versions currently use the item's created_at field to set the version's timestamp. 

This is wrong in the case where the item was created at some point in the past, then deleted, then is now being created again (e.g., through reification). The version gets the timestamp from the original creation.

Instead, this change uses the item's updated_at field which is the time of the re-creation. This is also consistent with what happens in "update" versions