Docs: It's not a breaking change, it's a fix :)
The "changes" section of the release notes is for breaking
changes only. The purpose is so that people don't need to read
as much when upgrading, they can just read the changes section,
they don't need to read the full release notes.

Only major version bumps should have breaking changes.

[ci skip]